Move service interface to common package
