+++ app/src/test/java/io/apicurio/registry/maven/TestUpdateRegistryMojoTest.java
@@ -17,7 +17,7 @@
 
 package io.apicurio.registry.maven;
 
-import io.apicurio.registry.client.RegistryService;
+import io.apicurio.registry.service.RegistryService;
 import io.apicurio.registry.rest.beans.ArtifactMetaData;
 import io.apicurio.registry.rest.beans.Rule;
 import io.apicurio.registry.types.ArtifactType;
