+++ client/src/main/java/io/apicurio/registry/client/CachedRegistryService.java
@@ -17,6 +17,7 @@
 package io.apicurio.registry.client;
 
 import io.apicurio.registry.rest.beans.*;
+import io.apicurio.registry.service.RegistryService;
 import io.apicurio.registry.types.ArtifactType;
 import io.apicurio.registry.types.RuleType;
 import io.apicurio.registry.utils.IoUtil;
