+++ tests/src/test/java/io/apicurio/tests/BaseIT.java
@@ -16,7 +16,7 @@
 
 package io.apicurio.tests;
 
-import io.apicurio.registry.client.RegistryService;
+import io.apicurio.registry.service.RegistryService;
 import io.apicurio.registry.rest.beans.ArtifactMetaData;
 import io.apicurio.registry.rest.beans.EditableMetaData;
 import io.apicurio.registry.types.ArtifactType;
