+++ utils/serde/src/main/java/io/apicurio/registry/utils/serde/AbstractKafkaDeserializer.java
@@ -16,7 +16,7 @@
 
 package io.apicurio.registry.utils.serde;
 
-import io.apicurio.registry.client.RegistryService;
+import io.apicurio.registry.service.RegistryService;
 import org.apache.kafka.common.serialization.Deserializer;
 
 import java.nio.ByteBuffer;
