+++ utils/serde/src/main/java/io/apicurio/registry/utils/serde/AbstractKafkaStrategyAwareSerDe.java
@@ -16,7 +16,7 @@
 
 package io.apicurio.registry.utils.serde;
 
-import io.apicurio.registry.client.RegistryService;
+import io.apicurio.registry.service.RegistryService;
 import io.apicurio.registry.utils.serde.strategy.ArtifactIdStrategy;
 import io.apicurio.registry.utils.serde.strategy.FindBySchemaIdStrategy;
 import io.apicurio.registry.utils.serde.strategy.GlobalIdStrategy;
