+++ utils/serde/src/main/java/io/apicurio/registry/utils/serde/ProtobufKafkaDeserializer.java
@@ -18,7 +18,7 @@ package io.apicurio.registry.utils.serde;
 
 import com.google.protobuf.Descriptors;
 import com.google.protobuf.DynamicMessage;
-import io.apicurio.registry.client.RegistryService;
+import io.apicurio.registry.service.RegistryService;
 import io.apicurio.registry.common.proto.Serde;
 
 import java.io.ByteArrayInputStream;
