+++ utils/tests/src/main/java/io/apicurio/registry/utils/tests/RegistryServiceExtension.java
@@ -17,7 +17,7 @@
 package io.apicurio.registry.utils.tests;
 
 import io.apicurio.registry.client.RegistryClient;
-import io.apicurio.registry.client.RegistryService;
+import io.apicurio.registry.service.RegistryService;
 import io.apicurio.registry.utils.IoUtil;
 import org.junit.jupiter.api.extension.*;
 import org.junit.platform.commons.util.AnnotationUtils;
